+++ lib/dash_kit_core.dart
@@ -6,18 +6,16 @@ export 'package:built_collection/built_collection.dart';
 export './src/components/action.dart';
 export './src/components/global_state.dart';
 export './src/components/store.dart';
-
 export './src/loadable/blocked_loadable_view.dart';
 export './src/loadable/loadable_grid_view.dart';
 export './src/loadable/loadable_item_view.dart';
 export './src/loadable/loadable_list_view.dart';
 export './src/loadable/loadable_paginated_grid_view.dart';
+export './src/loadable/loadable_paginated_list_both_direction_view.dart';
 export './src/loadable/loadable_paginated_list_view.dart';
 export './src/loadable/loadable_view.dart';
 export './src/loadable/pagination_state.dart';
-
 export './src/states/operation_state.dart';
 export './src/states/paginated_list.dart';
-
 export './src/utils/build_context_extensions.dart';
 export './src/utils/store_list.dart';