+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## [3.1.10]
+
+- Add new widget LoadablePaginatedListBothDirectionView
+- Fixed lists for using slivers
+
 ## [3.1.9]
 
 - Fixed lists for using slivers